Add support for the `enable_persistent_index` modification (#4958)
We add `enable_persistent_index` in properties to decide whether to use the persistent index in the PrimaryKey tablet. `enable_persistent_index` is specified when creating a table, and this pr adds support for the enable_persistent_index modification. We can use Alter table xxx set ("enable_persistent_index" = "false(true)") to modify it.